Started Comment implementation on Posts
This is a prototype for Comments on Posts. The list of comments is part of the post, but can be edited by anyone. Its implemented as an array of strings to keep it as simple as needed.
For now only one single comment can be added and they are not properly saved to the backend.